Rename reconcile.go to subscription.go
For consistency with Serving.
