+++ pkg/reconciler/v1alpha1/inmemory/controller/main.go
@@ -20,8 +20,8 @@ import (
 	"flag"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	"github.com/knative/eventing/pkg/reconciler/eventing/inmemory/channel"
-	"github.com/knative/eventing/pkg/reconciler/eventing/inmemory/clusterchannelprovisioner"
+	"github.com/knative/eventing/pkg/reconciler/v1alpha1/inmemory/channel"
+	"github.com/knative/eventing/pkg/reconciler/v1alpha1/inmemory/clusterchannelprovisioner"
 	"github.com/knative/eventing/pkg/provisioners"
 	istiov1alpha3 "github.com/knative/pkg/apis/istio/v1alpha3"
 	"github.com/knative/pkg/signals"
